permission added to admin users

+++ app/Http/Controllers/Admin/MasterEntry/AlliedBranchController.php
@@ -55,6 +55,9 @@ class AlliedBranchController extends Controller
      */
     public function create(formBuilder $formBuilder)
     {
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+
         $form=$formBuilder->create('nee_portal\Forms\AlliedBranchForm',
 
             ['method' =>'POST',
@@ -110,6 +113,9 @@ class AlliedBranchController extends Controller
      */
     public function edit($id, FormBuilder $formBuilder){
 
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+
             $allied_branch  = AlliedBranch::findOrFail($id);
 
                 $form    = $formBuilder->create('nee_portal\Forms\AlliedBranchForm',
@@ -158,6 +164,9 @@ class AlliedBranchController extends Controller
      */
     public function destroy($id)
     {
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+        
         AlliedBranch::destroy($id);
 
         if(Session::has('url')){
