+++ app/Http/Controllers/Admin/MasterEntry/CentreController.php
@@ -14,7 +14,7 @@ use Illuminate\Database\QueryException;
 
 use Kris\LaravelFormBuilder\FormBuilder;
 
-use Redirect;
+use Redirect, Basehelper;
 
 class CentreController extends Controller
 {
@@ -43,6 +43,9 @@ class CentreController extends Controller
      */
     public function create(formBuilder $formBuilder)
     {
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+
         $form=$formBuilder->create('nee_portal\Forms\CentreForm',
 
             ['method' =>'POST',
@@ -88,6 +91,9 @@ class CentreController extends Controller
      */
     public function edit($id, FormBuilder $formBuilder){
 
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+
             $centre  = Centre::where('centre_code', $id)->firstOrFail();
 
                 $form    = $formBuilder->create('nee_portal\Forms\CentreForm',
@@ -131,6 +137,9 @@ class CentreController extends Controller
      */
     public function destroy($id)
     {
+        if(Basehelper::Permission()==true)
+            return back()->with(array('message'=>'Access Denied!')); 
+        
         try{
 
               Centre::where('centre_code', $id)->delete();
